requirements.txt
django
django-compressor
django-ninja
gunicorn
pydantic[email]